Restructure to Service Oriented App Remove fgfApp Add Plants app Add Cultures app Add Animals app Update models for all apps
