+++ animals_app/views.py
