+++ auth_app/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class AuthAppConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'auth_app'
