+++ cultures_app/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CulturesAppConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'cultures_app'
